+++ OpenSim/Region/CoreModules/World/Terrain/FloodBrushes/RevertArea.cs
@@ -46,7 +46,7 @@ namespace OpenSim.Region.CoreModules.World.Terrain.FloodBrushes
         /// <param name="map">the current heightmap</param>
         /// <param name="fillArea">array indicating which sections of the map are to be reverted</param>
         /// <param name="strength">unused</param>
-        public void FloodEffect(ITerrainChannel map, bool[,] fillArea, double strength)
+        public void FloodEffect(ITerrainChannel map, bool[,] fillArea, float strength)
         {
             int x;
             for (x = 0; x < map.Width; x++)
